requirements.txt
python-gitlab>=3.15.0
python-dateutil>=2.8.2
tabulate>=0.9.0
